Photo_Step_Feature_fix
